#140 #143 renamed MoveableBoxComponent and made the old one deprecated. added documention to new code. fixed unit tests to stop failing with new component
