+++ source/core/src/main/com/csse3200/game/components/obstacles/MoveableBoxComponent.java
@@ -14,7 +14,13 @@ import com.csse3200.game.physics.components.ColliderComponent;
 import com.csse3200.game.physics.components.PhysicsComponent;
 import com.csse3200.game.rendering.TextureRenderComponent;
 
-public class MoveableBoxV2Component extends Component {
+/**
+ * This component is used to do the interactions between the player and the box entity. It is responsible for both
+ * interpreting the interaction when the player presses teh interact button and also for moving the host entity.
+ * It does this by taking the mouse pos in the world and slowly moving the box towards the mouse (within a max dist),
+ * the box also rotates based on the direction of where the mouse is.
+ */
+public class MoveableBoxComponent extends Component {
     private static final float INTERACT_RANGE = 1.5f;
     private static final float CARRY_RANGE = 1f;
 
@@ -28,9 +34,7 @@ public class MoveableBoxV2Component extends Component {
     private boolean pickedUp = false;
     private PhysicsComponent boxPhysics;
     private ColliderComponent boxCollider;
-    private ColliderComponent playerCollider;
     private TextureRenderComponent boxTexture;
-    private boolean playerInRange;
 
     private Camera camera;
     private final Vector2 tmp = new Vector2();
@@ -42,7 +46,14 @@ public class MoveableBoxV2Component extends Component {
     private boolean savedFixedRotation;
     private boolean savedBullet;
 
-    public MoveableBoxV2Component setCamera(Camera camera) {
+    /**
+     * Sets the internal camera variable.
+     * This camera variable is then used when doing mouse pointer calculations
+     *
+     * @param camera the world camera
+     * @return this component
+     */
+    public MoveableBoxComponent setCamera(Camera camera) {
         this.camera = camera;
         return this;
     }
@@ -68,21 +79,28 @@ public class MoveableBoxV2Component extends Component {
         boxPhysics.getBody().setGravityScale(BASE_GRAVITY_SCALE);
     }
 
+    /**
+     * Responsible for the initial interaction on the box.
+     * This has to be used to locate the player and then store it locally,
+     * it also sets up the required listener.
+     *
+     * @param collider the player collider
+     */
     public void setPlayerInRange(ColliderComponent collider) {
         if (collider == null) {
             return;
         }
 
-        playerInRange = true;
-        playerCollider = collider;
-
         if (!seenPlayer) {
             seenPlayer = true;
-            player = playerCollider.getEntity();
+            player = collider.getEntity();
             player.getEvents().addListener("interact", this::onPlayerInteract);
         }
     }
 
+    /**
+     * Toggle the physics filter so the box no longer interacts with the player
+     */
     private void toggleFilter() {
         Filter f = boxCollider.getFixture().getFilterData();
         if (pickedUp) {
@@ -97,6 +115,10 @@ public class MoveableBoxV2Component extends Component {
         boxCollider.getFixture().getBody().setAwake(true);
     }
 
+    /**
+     * What is run when a player presses the interact key.
+     * Only toggles the box lift if player is within the correct range.
+     */
     private void onPlayerInteract() {
         System.out.println("Box Interaction");
         if (player == null) {
@@ -110,10 +132,14 @@ public class MoveableBoxV2Component extends Component {
         }
     }
 
+    /**
+     * Toggle the box between the lifted and normal states.
+     * This includes toggling the physics filter, changing the gravity and
+     * various other physics related variables.
+     */
     private void toggleLift() {
         pickedUp = !pickedUp;
         Body body = boxPhysics.getBody();
-
 
         if (pickedUp) {
             // save
@@ -140,6 +166,8 @@ public class MoveableBoxV2Component extends Component {
 
     @Override
     public void update() {
+        // this has to be run here just once as the fixture body isn't initialised until
+        // after the create() method has been called.
         if (!appliedFilter) {
             if (boxCollider == null) return;
             if (boxCollider.getFixture() == null) return;
@@ -153,11 +181,18 @@ public class MoveableBoxV2Component extends Component {
             System.out.println("Filter applied");
         }
 
+        // run following method if box is currently picked up
         if (pickedUp) {
             followPlayer();
         }
     }
 
+    /**
+     * The core method driving the following functionality. It gets the mouse
+     * position in the world and then calculates a target pos for the box to float at
+     * which is scaled off of the {@code CARRY_RANGE}. It then applies a velocity to the box
+     * to push it towards the correct position.
+     */
     private void followPlayer() {
         if (player == null || camera == null) return;
 
@@ -181,11 +216,12 @@ public class MoveableBoxV2Component extends Component {
         float angle = MathUtils.atan2(dir.y, dir.x);
         boxTexture.setRotation((float) Math.toDegrees(angle));
 
+        // get delta away from target
         var body = boxPhysics.getBody();
         Vector2 center = body.getWorldCenter();
         Vector2 delta = tmp.cpy().sub(center);
 
-        // velocity proportional to error with max speed
+        // velocity proportional to delta with max speed
         float dist = delta.len2();
         Vector2 vel = (dist < 1e-4f) ? delta.setZero()
                                      : delta.scl(CARRY_GAIN);
@@ -193,6 +229,7 @@ public class MoveableBoxV2Component extends Component {
             vel.setLength(CARRY_MAX_SPEED);
         }
 
+        // apply correct velocity and don't rotate at all
         body.setAngularVelocity(0f);
         body.setLinearVelocity(vel);
 
@@ -227,7 +264,4 @@ public class MoveableBoxV2Component extends Component {
         entity.setPosition(tmp);
         boxTexture.setRotation((float) Math.toDegrees(angle));
     }
-
-
-
 }
